update: series bar refactor
